ISSUE #2943 - move autocomplete options to their own files
